requirements.txt
streamlit==1.35.0
scikit-learn==1.5.0
nltk==3.8.1
pandas==2.2.2
numpy==1.26.4
regex==2024.5.15
openai
python-dotenv
google-generativeai==0.5.4
